Revert "CameraControllers: use the Camera's upVector"
This reverts commit 627167bd4a2e1dc12ffee813e82f9b33bff19a22.

There was a reason to use the y axis as the up vector here which
is to keep the camera controller oriented correctly. Without this
restriction it's far too easy to end up with a twisted camera.

If we want a fully free trackball type controller, let's add that
as a new type, or perhaps as an option on these.

Change-Id: I56c8708535d4ccd05ccb5f2ddea732bb27abda58
Reviewed-by: Mike Krus <mike.krus@kdab.com>