Don't assume position attribute uses default name in visitPrimitives
It is not ok to assume that the position attribute has the default
position attribute. It should be the fallback behavior like it is in
calculateLocalBoundingVolume. In this case we should try and query the
boundingPositionAttribute if it is set.  If anything this could have
been called just "positionAttribute" since we need to introspect mesh
data in a few places and need to know the positionAttribute.

The previous logic meant that meshes used by Qt 3D Studio could not take
advantage of advanced picking (anything other than bounding spheres)
because it does not use the built-in materials or default attribute
names (they are optional after all).

Change-Id: I5b02a0532add1582ae48121cdc6a7b3040ad1cdd
Reviewed-by: Mike Krus <mike.krus@kdab.com>