Fix race in startup/shutdown
When the AspectEngine with a QRenderAspect (and RenderThread) is started and
stopped within the same loop, there was a race in which Renderer::shutdown
would possibly be called while Renderer::initialize wasn't complete.

Apart from testing, such starting and stopping within the same loop is
strongly discouraged as Qt3D doesn't have time to create an offscreen surface
to properly release GL resources.

Change-Id: Ie6f0a0698c213e3015bc9c51efc48d8706f20b3b
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>