Don't abort rendering when env light textures are not set
The MetalRoughMaterial provides parameters for the env light textures even
though they are optional but are needed to avoid issues with unbound textures
on some drivers.
In practice, when an env light is present, the values would be overridden.
However if no env light is present, when setting the texture values on the
shader, they would actually reference invalid textures and Qt3D would not
render. We want to keep the behavior for the general case but make an
exception for MetalRoughMaterial, therefore we are checking if the missing
textures are the ones for env light or not before aborting.

Ideally, we should shortly be able to remove the QParameters altogether with
the Shader Builder.

Change-Id: Id54436ff9e328c390eacceb3e63809c757b28b2a
Reviewed-by: Kevin Ottens <kevin.ottens@kdab.com>