Automatically register the latest import version
This follows what was done in d28c9f6a for Qt Quick Controls 2.
The latest import version (e.g. 2.11 in Qt 5.11) will automatically be
registered whenever the Qt version is bumped.

This avoids needing to wait until a new type is added (or a new
revision is added to an existing type) before being able to use the
newest Qt Quick version.

Task-number: QTBUG-70290
Change-Id: Id221b5879f8c50a20fe5e26f9e141b82e4a196cb
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>