Fix crash due to racecondition in renderer on shutdown
Protects access to m_renderQueue using the m_renderQueue mutex on
shutdown. When a Qt3DWindow is deleted it signals the aspect thread to
do early cleanup, the aspect thread calls Renderer::shutdown which in
turn resets m_renderQueue. Meanwhile the render thread may be working on
m_renderQueue, this leads to a double free of the front render view in
m_renderQueue (and potentially other stuff).

Fixes: QTBUG-69877
Change-Id: Ifd018597dbdb2f1538494932c907c63e8db9c715
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>