QRenderCapture: Prevent crash on shutdown when replies are pending
When a QRenderCapture is deleted its destructor runs, then the
QFrameGraphNode destructor, then the QNode destructor and finally the
QObject destructor. The QObject destructor deletes its QObjectPrivate
member, which destructs QRenderCapturePrivate. At that point the
QRenderCapturePrivate tries to disconnect on an invalid pointer to the
QRenderCapture (which was already deleted).

Change-Id: Ib5835e3f86a282b963220b0dab656258111a61ab
Reviewed-by: Antti Määttä <antti.maatta@qt.io>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>