Fix MouseEvent/WheelEvent to properly forward all keyboard modifiers
Modifiers was an enum when it should have been a flag.
This patch makes the transition toward having a flag.
The Modifiers enum is renamed Modifier and the flag
is named Modifiers.

This might affect BIC but this change is really
needed to make the modifiers property of Mouse/Wheel
event actually useful.

Change-Id: I67e472c3cc2654f8d3e84c6248b8e689c0306530
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>