Rework internal Texture Handling
- Remove internal GLTexture sharing
- Remove dataFunctor data sharing
- Simplify the overall texture process:
  1) Look for dirty textures
     (either a property on QTexture or a referenced QTextureImage
     has changed)
  2) Update GL Resources
     - Create a GLTexture for a given Texture if it doesn't exist
     - Update GLTexture to reflect Texture state
     - Perform actual GL texture creation and or upload
       (call functors at this point)
     - Cleanup GLTextures when matching Texture node has been
       destroyed

This will provide an easier maintenance over time as it drastically
simplifies the handling of textures and removes most of the coupling.
Furthermore this will make it possible to send texture updates on
a QTexture node to update texture content without having to declare
QTextureImages which would not have been (or would have been a lot
harder) to do. In practice, I doubt most people were even aware of
the internal texture sharing in the first place.

[ChangeLog][Qt3DRender] Textures: internal data sharing removed

Change-Id: I02867c8105e29eb1e193884e3899062f795f32f4
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>