Always load scene2d render plugin in static builds
Scene2d requires support from the Qt3D renderer, which is implemented
in form of a renderer plugin. The renderer plugin is implicitly loaded
when the scene2d module is loaded, through use of
Q_COREAPP_STARTUP_FUNCTION, which does not support static builds,
though.
This change adds code to initialize the scene2d render plugin to the
constructor of QScene2D, so that the first QScene2D object created will
initialize the renderer plugin.

Task-number: QTBUG-72378
Change-Id: I6f625e79a5be9f5a23dcc881f9b360382008032b
Reviewed-by: Janne Koskinen <janne.p.koskinen@qt.io>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>