Scene3D: Call update on the item not the window
On scene change QQuickWindow::update was called to ask the QQ2Renderer
to rerender the scene. The problem here is, that calling
QQuickWindow::update is percieved as a call to rerender under all
circumstances, while QQuickItem::update only means update if the scene
graph actually changed in the end. Using QQuickWindow::update the scene
will be rerendered even if the Scene3D item is invisible, using
QQuickItem::update it won't.

Change-Id: I1ff455628d3e536349683b064acaeadd354ec864
Fixes: QTBUG-72923
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>