Update Qt3D color handling
There have been a number of updates to QtQuick color handing that were
not similarly updated on the Qt3D side (for example the addition of
hsl* and hsv* properties). This led to different behavior depending on
the imports used, where the Qt3D behavior did not match the basic color
documentation.

[ChangeLog][Quick3D] Update Qt3D color handing to match QtQuick.

Task-number: QTBUG-75551
Change-Id: I13d163c2ada39dd8941351cb0cfa991618333ee1
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>