Fix following recent Scene3DRenderer changes
f39178a415cb41470775a86e0aa358faa3686d81 makes QtQuick redraw only when
something in the 3D scene has triggered a change.
A function is connected to QQuickWindow::afterSynchronizing to check
whether the 3D scene needs to be drawn again and if so the corresponding
SceneGraph nodes are marked as dirty. However, we need to ensure the SceneGraph
picks up that we've marked a node as dirty by calling QQuickItem::update().
This forces updatePaintNode to be called at the next frame since the
afterSynchronizing stage occurs after that step.

This fixes rendering freezes with the Always render policy.

Change-Id: I041dbe69d12a7e6e884c8ee0e29720d74ce5fa2f
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>