Doc: Update code samples in 'Qt 3D Render Framegraph' documentation
The code snippets were outdated, referring to QML properties that
do not exist. In particular, the Deferred Rendering section contains
code that was based on an example that's since moved into tests.
Adapt the code to match the updates done to that example and also
mention where it originates from.

Task-number: QTBUG-73024
Change-Id: I9ec750d05f8e853e5614f33dadcbe44a0b42e356
Reviewed-by: Paul Wicking <paul.wicking@qt.io>