QEventFilterService: these aren't the algorithms you're looking for
The event filters are ordered by priority, so don't use a linear scan
in order to find whether one with the same position already exists,
use lower_bound, whose result at the same time gives the position of
an existing filter (if any), as well as the insertion position for the
new element. So no need to do a full sort afterwards, either.

Change the exisitng op< into an aptly-named lambda, and don't depend on
transitive includes.

Also, since this code does not need the CoW feature of QVector, port to
std::vector, which is faster and produces more compact code.

Altogether saves ~2KiB (0.47%) in text size on optimized Linux AMD64
GCC 9.1 builds.

Change-Id: Ie46ca2fbb2dcbea23a0098a72336604776dcc296
Reviewed-by: Mike Krus <mike.krus@kdab.com>