QImageTextureDataFunctor: return invalid data when url is invalid
Since QSkyboxEntity creating empty cube map first and using QTimer to
delay loading real cube map, in render thread, GLTexture will use the
generator function of empty cube map to generate a QTextureImageDataPtr
at first, this will set the format of property of GLTexture to NoFormat,
and terminate the creation of GLTexture. Although GLTexture read real
real cube map next time, format will not be reset and the creation of
GLTexture terminates.

[ChangeLog][Qt3DRender][QImageTextureDataFunctor] return a invalid
image data when url is invalid to ensure the property of GLTexture will
not be set to NoFormat

Fixes: QTBUG-74017
Change-Id: If48f727dba817a931d618b2b88c7ebd54c508c47
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>