Add a property map job in animation aspect
ClipAnimator is always dirty because at each frame the normalizedTime is
changed. This has a big performance penalisation as we are launching
 job which updates the property map at each frame.

This patch adds a new dirty flag which is use to decide if the animator
need to update the mapping of the fcurve to actual node properties.

Using that flag, a new UpdatePropertyMapJob is launched for dirty
clipAnimators.

Some scenes boost from 30fps to 60fps (vsync) during animation.

Change-Id: Ic8a8db01535c55995bc569ea5a69660a40014401
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>