RenderSettings: only destroy the RenderSettings component that is used
Since we only create a backend node for the first RenderSettings component that
is inserted in the scene, we should only destroy that instance. Otherwise, in
the case we have several QRenderSettings components in a scene, if one of them
gets destroyed we might end up destroying the RenderSettings instanced actually
used by the Renderer.

We so far assumed there would only be a single QRenderSettings per application
but introducing Scene3DView might change that.

Change-Id: Ibef1328253b871f3b40a2298ccd8504130d75d86
Reviewed-by: Mike Krus <mike.krus@kdab.com>