QDistanceFieldGlyphCache: make sure the arbiter of atlas can be set
When using QText2DEntity to display text, the node of QTextureAtlas
will be create when traversing the node tree, and it will use root
node as its parent node. However, since the children node of root
node is being traversed in this moment, QTextureAtlas will not be
traversed, that means the change arbiter won't be set, and the change
in frontend cannot notify backend.

[ChangeLog][Qt3DExtras][QDistanceFieldGlyphCache] make sure the change
arbiter of atlas for QText2DEntity can be set when traversing node tree,
and add new auto test for this.

Fixes: QTBUG-77084
Change-Id: Iad09efe4376ba61f54624965118358a160be42a2
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>