Fix crash in tst_QShaderProgram
We install the TestArbiter on "this" in one test and the TestArbiter is
destroyed at the end of that test. In a second test we again install a
TestArbiter on "this".  The call to QNode::setArbiter tries to access
the desroyed arbiter (m_changeArbiter->removeDirtyFrontEndNode(q))
resulting in a crash.

Need to set a null arbiter on "this" to properly clean up.

Also fix a few other incorrect uses that weren't crashing.

Change-Id: Iee2ea8210df971f934e2f188748803c8745c8278
Reviewed-by: Mike Krus <mike.krus@kdab.com>