Remove backend entity hierarchy rebuilding jobs
We no longer need to have a special job to rebuild the entity
hierarchy as parent nodes can now always be resolved on the
backend either at creation time or while reparenting.

The NodePostConstructorInit ensure that all backend nodes are
created from the top down so a node can always find it's
parent.

Change-Id: I9dcfb64b62e2e02b89b1ae59efb41f25dfc09eae
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>