Fix/improve several node tests
Changing the backend to remove syncing removed creation
and destruction events and the tests were updated to reflect
that. The key part of several tests actually involve this
creation and destruction information, such as testing to ensure
all parent backend nodes are created before their children.

Created a TestAspect that allows capturing the creation/destruction
of backend nodes and updated the many of the tests to check this
information

These tests will be important when refactoring to remove the
entity hierarchy rebuilding and some other cleanup.

Change-Id: I1e6a7ce01df2c711b266bccc6fe3dcb31431e3f1
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>