Fix test for parenting QEntity to QNode
The purpose of this test is to make sure when parenting a QEntity to a
QNode, the backend entity will receive notification and be able to
correctly resolve it's parent entity.  Update the test to check that the
QEntity node is marked dirty in the arbiter and has the correct parent
and parent entity.

Change-Id: I8f2c8aa1c8b5b6bbabd956bbfdef36fafcd05f8d
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>