PackUniformHash to QVector
Surprisingly it's hard to notice its effect in the speed of execution,
frame preparation looks to be more or less the same with the profiler.
However with vtune, the profiling traces show a huge difference with
QHash, mainly in time spent allocating memory. It shows a noticeable
reduction in CPU usage.

On bigscene-cpp with 600 entities

QHash   -> On a 158s run, CPU time is 112s (70%)
           free accounts for 26s (23%), malloc 24s (21%)

QVector -> On a 190s run, CPU time is 110s (58%)
	   free accounts for 5s (4.5%), malloc 4.7s (4.2%)

Change-Id: I880d44b1acf7f051e479ed356864c3caf407f23f
Reviewed-by: Mike Krus <mike.krus@kdab.com>