Temporarily disable binary JSON mode for qgltf in bootstrap mode
We need to update the bootstrap library in order to remove binary JSON
support and add CBOR support. As we cannot use CBOR before and we cannot
use binary JSON afterwards, we need to disable binary JSON in qgltf,
then update the bootstrap library, and then add CBOR support to qgltf.

Change-Id: I4d70b831cf1924d08c821c74f0c1b63fd0d21844
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>