Fix status change on QTextureLoader
Now setting to Loading when the functor change is synced to the backend.
Avoid creating a new functor when downloads have completed, just use
the same one and mark the node as dirty.
Avoid recreating the functor when the format change comes from the
backend (which caused the image to be loaded again).

Change-Id: Ifcbbf54db56b57a58b0b5d1ce5f1475b6587f697
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>