Restore default rendering behavior of Scene3D
Recent change in Scene3D was made to stop it rendering when it
was hidden. However this means it no longer can be used as a hidden
item in combination with an OpacityMask.

Revert default behavior to always draw and introduce an environment
variable to let the user stop rendering.

5.15 will add a property to control this properly.

Change-Id: I30f95b2704cc7c1c0a4467677c8a9e4e1dbb9962
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>